Quick fix for automatic mode
Currently a peerCount of -1 means we couldn't connect to the external ipfs proc.
This updates the automatic gateway switching logic to use that signalling value
to decide whether to switch to the public gateway. Zero peers means we're
initialising, rather than we'er offline.

This improves the ux when switching between embedded and external nodes.

A more extensive change can be done later, wrapping the ipfs-api client with a
cache, and emitting events for when ipfs state changes.