fix: restore gateway redirect for CORS XHR in Firefox
Context for CORS XHR problems in Firefox:
https://github.com/ipfs-shipyard/ipfs-companion/issues/436

In short, onBeforeRequest should not change anything, as it will trigger
false-positive CORS error.  onHeadersReceived is after CORS validation
happens, so its ok to cancel and redirect late.

This is not ideal, as there is an outgoing request to the public
gateway, and we need to read response headers before connection is
aborted, but we can't do better than that until
https://bugzilla.mozilla.org/show_bug.cgi?id=1450965 is resolved.