feat: add libdweb-build target
This is a first step on the road to native protocol handler and more.

mozilla/libdweb is included as a git submodule, which enables us to
develop against latest API prototypes with as little overhead as
possible.

`libdweb-build` will produce extension bundle with access to
experimental APIs

`firefox:libdweb` will start Firefox Nightly with mentioned bundle.
Make sure firefox-nightly is on your PATH