fix: disable redirect for CID-in-subdomain
This is a fix of a regression for
https://github.com/ipfs-shipyard/ipfs-companion/pull/537
that was introduced when X-Ipfs-Path header support was added.

Until we are able to provide the same origin-based security guarantees
at a local gateway, we should not redirect resources that use
cid-in-subdomain deployment, as it is a strong hint they care about
Origin-based isolation and we should not relax those guarantees.