feat: preload webui root CID
Optimization: preload the root CID to speed up the first time
Web UI is opened. If embedded js-ipfs is used it will trigger
remote (always recursive) preload of entire DAG to one of preload nodes.
This way when embedded node wants to load resource related to webui
it will get it fast from preload nodes.

This is a best-effort version of #682