feat: opt-in for /ipns/webui.ipfs.io
Adds an opt-in toggle (disabled by default) to Preferences
which changes the URL of Web UI opened via Browser Action menu
from {API}/webui to {API}/ipns/webui.ipfs.io

This enables user to load the latest webui via DNSLink.

Note that go-ipfs and js-ipfs do not whitelist /ipns/webui.ipfs.io on
the API port yet, so there is a fallback in place that detects HTTP 404
and redirects user to {API}/webui.

Closes: https://github.com/ipfs-shipyard/ipfs-companion/issues/736