feat: remove 'redirect on {domain}' from main menu
This removes "Redirect on {fqdn}" menu item from browser action menu.
DNSLink redirect is no longer enabled by default, and this feature does
not need to be present at all times (It remains on Preferences screen if
needed).

In the future we will change it to "Disaple IPFS integrations on these
hostnames"

Part of #741