fix: remove normalizeLinksContentScript (#836)
The script was responsible for overriding all src and href in DOM of
every page to ensure content addressed with ipfs:// URL loads.

Looking back, it does not align well with our upgrade path, and the
result across browser vendors was buggy. While it allowed loading simple
images, it did not help with broken XHR and iframe quirks and introduced
unnecessary battery drain.

Due to this, it is better to remove it and advise users to use
content-addressing at a public gateway.
This way everything works for people without IPFS client, and browsers
with IPFS support can easily detect it and upgrade the call.