fix: quick upload in Chromium
In short, there is a discrepancy between how objects are passed between
a page and background page within the same extension in Firefox and
Chromium.

Quick import in Chromium was broken, most likely by additional process
boundary serialization that was not present in Firefox, could be related
to upstream bug at:
https://bugs.chromium.org/p/chromium/issues/detail?id=112163

Potential fix via trampoline from https://stackoverflow.com/a/21101597/11518426
is too brittle, as Chromium can break it in browser context at any time.

It seems that ReadableStream produced by file.stream() is able to
correctly pass the boundary between quick import page and the IPFS API
object running in the background, so we use it with embedded js-ipfs,
but for external node we use http client directly from quick upload
page, and pass File object as-is, which takes 1/2 of time.

Additional benchmarks can be found in
https://github.com/ipfs-shipyard/ipfs-companion/pull/811