refactor: remove webui TAR from Chromium bundle
We were bundling TAR inside of Chromium package to make it work
instantly when opened in Brave.

Unfortunately it added 10MB to the package, which meant one-click
install was slow in Brave, but also on every other Chromium browser.

This replaces bundled TAR with prefetching it from a public gateway on
the first run. The idea is that if user had access to Chrome Web Store
to install extension, they have access to the gateway, so no need for
bundling.

We also precache regular webui in Chromium and Firefox via ipfs.refs