fix: redirect to native URI in Brave
This is a cosmetic fix that ensures IPFS resources are redirected to
ipfs:// or ipns:// URI and that URI is displayed in the browser UI
in the address bar.

Due to the way Brave implemented URI cloak for local HTTP-based gateway
provided by go-ipfs managed by the browser itself the regular redirect
done via webRequest API did not produce native URI in address bar.

However, tabs.update does the trick, so we use that for root requests.