feat: replace pinning with import to mfs
This changes the way IPFS content is kept around.
Instead of using low level pins we copy the item to MFS,
unifying experience to match recent changes in ipfs-webui v2.12
and ipfs-desktop v0.15.0

This also removed page-action, because it was Firefox-specific feature
and made maintenance and testing more difficult
(now we have same UX in all browsers).

Closes #742
Closes #888
Closes https://github.com/ipfs/ipfs-gui/issues/91